+++ server/src/main/kotlin/com/lucasalfare/flpoint/server/c_infra/webserver/ktor/configuration/RoutingConfiguration.kt
@@ -2,8 +2,11 @@ package com.lucasalfare.flpoint.server.c_infra.webserver.ktor.configuration
 
 import com.lucasalfare.flpoint.server.b_usecase.PointUsecases
 import com.lucasalfare.flpoint.server.b_usecase.UserUsecases
-import com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes.*
 import com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes.admin.adminPointsCrudRoutes
+import com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes.dummy.protectedForAdminRoute
+import com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes.dummy.protectedRoute
+import com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes.point.pointsHandlingRoutes
+import com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes.user.userAccessRoutes
 import io.ktor.server.application.*
 import io.ktor.server.routing.*
 
@@ -14,10 +17,8 @@ fun Application.routingConfiguration(
   routing {
     protectedRoute()
     protectedForAdminRoute()
-    registerPostRoute(userUsecases)
-    loginPostRoute(userUsecases)
-    pointRoute(pointUsecases)
-    listPointsRoute(pointUsecases)
+    userAccessRoutes(userUsecases)
+    pointsHandlingRoutes(pointUsecases)
     adminPointsCrudRoutes(pointUsecases)
   }
 }
