+++ server/src/main/kotlin/com/lucasalfare/flpoint/server/c_infra/webserver/ktor/routes/point/PointsHandlingRoutes.kt
@@ -1,4 +1,4 @@
-package com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes
+package com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes.point
 
 import com.lucasalfare.flpoint.server.a_domain.model.dto.CreatePointRequestDTO
 import com.lucasalfare.flpoint.server.b_usecase.PointUsecases
@@ -10,7 +10,7 @@ import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
-fun Routing.pointRoute(pointUsecases: PointUsecases) {
+fun Routing.pointsHandlingRoutes(pointUsecases: PointUsecases) {
   authenticate("flpoint-jwt-auth") {
     post("/point") {
       val principal = call.principal<JWTPrincipal>()
@@ -18,6 +18,12 @@ fun Routing.pointRoute(pointUsecases: PointUsecases) {
       val req = call.receive<CreatePointRequestDTO>()
       val result = pointUsecases.createTimeRegistration(userId, req)
       return@post call.respond(HttpStatusCode.Created, result.getOrNull()!!)
+    }
+
+    get("/points") {
+      val userId = call.principal<JWTPrincipal>()?.payload?.getClaim("id")?.asInt()!!
+      val result = pointUsecases.getAllUserPoints(userId)
+      return@get call.respond(HttpStatusCode.OK, result)
     }
   }
 }
