+++ server/src/main/kotlin/com/lucasalfare/flpoint/server/c_infra/webserver/ktor/routes/user/UserAccessRoutes.kt
@@ -1,6 +1,7 @@
-package com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes
+package com.lucasalfare.flpoint.server.c_infra.webserver.ktor.routes.user
 
 import com.lucasalfare.flpoint.server.a_domain.model.dto.BasicCredentialsDTO
+import com.lucasalfare.flpoint.server.a_domain.model.dto.CreateUserDTO
 import com.lucasalfare.flpoint.server.b_usecase.UserUsecases
 import io.ktor.http.*
 import io.ktor.server.application.*
@@ -8,11 +9,17 @@ import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
-// curl -v -H 'Content-Type: application/json' -d '{"email": "asdf@qwer.com", "plainPassword": "zxcv"}' -X POST http://localhost:7171/login
-fun Routing.loginPostRoute(userUsecases: UserUsecases) {
+// routes that doesn't requires authentication to access
+fun Routing.userAccessRoutes(userUsecases: UserUsecases) {
   post("/login") {
     val dto = call.receive<BasicCredentialsDTO>()
     val jwt = userUsecases.loginUser(dto)
     call.respond(HttpStatusCode.OK, jwt)
+  }
+
+  post("/register") {
+    val dto = call.receive<CreateUserDTO>()
+    val result = userUsecases.createUser(dto)
+    call.respond(HttpStatusCode.Created, result.getOrNull()!!)
   }
 }
